Create a Reusable, AST-Based Build-Time Processor #7158
